refactor(ipsec): split policy.go, lift address-resolution island
pkg/ipsec/policy.go had grown to 1,135 LOC (+29% since the last flag)
and fused two unrelated aspects: swanctl config rendering and the
config-prep + local address/family-resolution machinery. This is a pure
code-motion split — no logic edit, rename, or reorder — mirroring the
recent #6441/#6442/#6443 sibling splits.

The bottom ~490 LOC is a self-contained island: PrepareConfig and the
address/family-resolution helpers (HasDHCPBoundGateway,
interfaceRefIsDHCP, resolveInterfaceAddress[Family], the resolveHostFamily
machinery, resolveGatewayFamilyHints, gatewayRemoteFamilyHint,
resolveConfiguredInterfaceAddress, selectUnitAddress,
resolveKernelInterfaceAddress, selectFamilyAddress, bareIP,
bareIPGlobalOnly, parseBareIP, zoneQualify, addressFamilyHint,
matchFamily). It has no upward deps and nothing in the render half calls
into it, so it lifts out verbatim to a new sibling pkg/ipsec/policy_addr.go.
policy.go keeps the swanctl render aspect and drops to 637 LOC;
policy_addr.go is 511 LOC.

Only the per-file package/import lines differ. The import split is
mutually exclusive except the shared sort/strings/config: policy.go drops
context/net/strconv/sync/time (used only by the moved island),
policy_addr.go gains them.

Behavior-preservation proof:
  norm() strips package/import/blank/comment-only lines.
  git show origin/master:pkg/ipsec/policy.go | norm | sort  -> 697 lines
  cat policy.go policy_addr.go            | norm | sort  -> 697 lines
  diff of the two multisets                             -> empty (zero delta)
  ordered byte-diff old lines 639-1135 vs the moved decls -> empty (verbatim)

Validation: go build ./..., go vet ./pkg/ipsec/, go test ./pkg/ipsec/,
and the FULL go test ./... all GREEN, including the #1373 pkg/dataplane
retirement canary and pkg/refactoraudit TestHeatmapNotStale. policy.go
was 1,135 LOC (below the 1,500 restructure trigger), so no heatmap
regeneration was required.

Closes #6439
